requirements.txt
apache-airflow==2.10.4
apache-airflow-providers-postgres
apache-airflow-providers-docker
apache-airflow-providers-apache-kafka
apache-airflow-providers-apache-spark==5.2.1
apache-airflow-providers-smtp
pandas
psycopg[binary]
pyspark==3.5.1
kafka-python==2.0.2
confluent-kafka==2.3.0
black==24.3.0
pylint==3.1.0
pre-commit==3.7.0
pytest==8.1.1
pytest-cov==5.0.0
pytest-mock==3.14.0
